requirements.txt
pandas
google-api-python-client
tqdm
emoji
langdetect
googletrans==4.0.0-rc1
tensorflow==2.20.0
keras
scikit-learn
joblib
streamlit
numpy
scikit-learn
nltk
matplotlib
seaborn
plotly
regex
python-dotenv
streamlit
pandas
numpy
plotly
matplotlib
wordcloud
openpyxl
XlsxWriter
python-dateutil
reportlab
malaya

